Refactor for Vercel
