+++ .gitignore
@@ -17,3 +17,4 @@
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
+.vercel
