--- api/package.json
@@ -1,12 +1,0 @@
-{
-  "name": "marriage-index-api",
-  "version": "1.0.0",
-  "scripts": {
-    "start": "func start"
-  },
-  "dependencies": {
-    "azure-storage": "^2.10.6",
-    "azure-table-storage-async": "^1.3.0",
-    "dotenv": "^10.0.0"
-  }
-}
