+++ src/Search.jsx
@@ -12,10 +12,9 @@ export default function Search() {
 
   const submitForm = async (e) => {
     e.preventDefault();
-    const url = process.env.REACT_APP_SEARCHURL;
     try {
       setLoading(true);
-      const response = await fetch(url, {
+      const response = await fetch("/api/search", {
         method: "POST",
         headers: {
           "Content-Type": "application/json",
@@ -58,88 +57,88 @@ export default function Search() {
 
   return (
     <div>
-      {loading ? (
-        <h1>Loading...</h1>
-      ) : (
-        <Parallax
-          image={<img alt="" src="/images/pexels-photo-130111.jpeg" />}
-          options={{
-            responsiveThreshold: 0,
-          }}
-        >
-          <section>
-            <div className="container search-form white" style={{ padding: "3px" }}>
-              <h4 className="brown-text center-align">Search the database</h4>
-              <h6 className="brown-text center-align" style={{ marginBottom: "25px" }}>
-                Please enter a surname as a minimum. A First name or Year of marriage can be considered as
-                optional
-              </h6>
+      <Parallax
+        image={<img alt="" src="/images/pexels-photo-130111.jpeg" />}
+        options={{
+          responsiveThreshold: 0,
+        }}
+      >
+        <section>
+          <div className="container search-form white" style={{ padding: "3px" }}>
+            <h4 className="brown-text center-align">Search the database</h4>
+            <h6 className="brown-text center-align" style={{ marginBottom: "25px" }}>
+              Please enter a surname as a minimum. A First name or Year of marriage can be considered as
+              optional
+            </h6>
 
-              <div className="row">
-                <form className="col s10 offset-s1" onSubmit={submitForm}>
-                  <div className="row">
-                    <div className="input-field col s12">
-                      <i className="brown-text material-icons prefix">account_circle</i>
-                      <input
-                        id="surname"
-                        type="text"
-                        className="validate"
-                        name="surname"
-                        required
-                        onChange={(e) => setSurname(e.target.value)}
-                      />
-                      <label htmlFor="surname">Surname</label>
-                    </div>
-                    <div className="input-field col s6">
-                      <i className="brown-text material-icons prefix">face</i>
-                      <input
-                        id="firstname"
-                        type="text"
-                        className="validate"
-                        name="firstname"
-                        onChange={(e) => setFirstName(e.target.value)}
-                      />
-                      <label htmlFor="firstname">First Name</label>
-                    </div>
-                    <div className="input-field col s6">
-                      <i className="brown-text material-icons prefix">date_range</i>
-                      <input
-                        id="year"
-                        type="text"
-                        className="datepicker"
-                        name="year"
-                        onChange={(e) => setYear(e.target.value)}
-                      />
-                      <label htmlFor="year">Year</label>
-                    </div>
+            <div className="row">
+              <form className="col s10 offset-s1" onSubmit={submitForm}>
+                <div className="row">
+                  <div className="input-field col s12">
+                    <i className="brown-text material-icons prefix">account_circle</i>
+                    <input
+                      id="surname"
+                      type="text"
+                      className="validate"
+                      name="surname"
+                      required
+                      onChange={(e) => setSurname(e.target.value)}
+                    />
+                    <label htmlFor="surname">Surname</label>
                   </div>
-                  <button className="btn waves-effect waves-light right" id="search-submit" type="submit">
-                    Submit
-                    <i className="material-icons right">search</i>
-                  </button>
-                </form>
-              </div>
+                  <div className="input-field col s6">
+                    <i className="brown-text material-icons prefix">face</i>
+                    <input
+                      id="firstname"
+                      type="text"
+                      className="validate"
+                      name="firstname"
+                      onChange={(e) => setFirstName(e.target.value)}
+                    />
+                    <label htmlFor="firstname">First Name</label>
+                  </div>
+                  <div className="input-field col s6">
+                    <i className="brown-text material-icons prefix">date_range</i>
+                    <input
+                      id="year"
+                      type="text"
+                      className="datepicker"
+                      name="year"
+                      onChange={(e) => setYear(e.target.value)}
+                    />
+                    <label htmlFor="year">Year</label>
+                  </div>
+                </div>
+                <button className="btn waves-effect waves-light right" id="search-submit" type="submit">
+                  Submit
+                  <i className="material-icons right">search</i>
+                </button>
+              </form>
             </div>
-          </section>
-        </Parallax>
-      )}
+          </div>
+        </section>
+      </Parallax>
 
-      <div className="section brown ">
-        <div>
-          <div className="row">
-            <div className="col s10 offset-s1">
-              <Results data={data} />
+      {data ? (
+        <div className="section brown ">
+          <div>
+            <div className="row">
+              <div className="col s10 offset-s1">
+                <Results data={data} />
+              </div>
             </div>
           </div>
         </div>
-      </div>
+      ) : (
+        ""
+      )}
     </div>
   );
 }
 
 const Loading = () => {
   return (
-    <div className="center-align" style={{ margin: "20vh" }}>
+    <div className="center-align" style={{ height: "80vh", paddingTop: "200px" }}>
       <Preloader active color="green" flashing={false} size="big" />
     </div>
   );
